Unify p2p sync and portal sync
