+++ kolibri/core/error_constants.py
@@ -11,10 +11,12 @@ INVALID = "INVALID"
 UNIQUE = "UNIQUE"
 INVALID_NETWORK_LOCATION_FORMAT = "INVALID_NETWORK_LOCATION_FORMAT"
 NETWORK_LOCATION_NOT_FOUND = "NETWORK_LOCATION_NOT_FOUND"
+INVALID_TOKEN = "INVALID_TOKEN"
 # 401 error constants
 INVALID_CREDENTIALS = "INVALID_CREDENTIALS"
 # 404 error constants
 NOT_FOUND = "NOT_FOUND"
+FACILITY_DOES_NOT_EXIST = "FACILITY_DOES_NOT_EXIST"
 # 403 error constants
 PERMISSION_DENIED = "PERMISSION_DENIED"
 NOT_AUTHENTICATED = "NOT_AUTHENTICATED"
