+++ kolibri/core/utils/portal.py
@@ -1,10 +1,8 @@
 import json
-import socket
 
 import requests
 from morango.api.serializers import CertificateSerializer
 from morango.certificates import Certificate
-from morango.certificates import Nonce
 from morango.constants.api_urls import NONCE
 from six.moves.urllib.parse import urljoin
 
@@ -12,12 +10,7 @@ from kolibri.core.auth.constants.morango_scope_definitions import FULL_FACILITY
 from kolibri.utils import conf
 
 
-def claim(token, facility, client_ip=None):
-
-    # get client ip if not passed in
-    if not client_ip:
-        hostname = socket.gethostname()
-        client_ip = socket.gethostbyname(hostname)
+def registerfacility(token, facility):
 
     # request the server for a one-time-use nonce
     PORTAL_URL = conf.OPTIONS["Urls"]["DATA_PORTAL_SYNCING_BASE_URL"]
@@ -35,11 +28,9 @@ def claim(token, facility, client_ip=None):
     if client_cert is None:
         raise Certificate.DoesNotExist
 
-    client_nonce = Nonce.objects.create(ip=client_ip)
     # build up data for request
     data = {
         "facility_id": facility.id,
-        "client_nonce": client_nonce.id,
         "server_nonce": server_nonce,
         "client_certificate_id": client_cert.id,
         "certificate_chain": json.dumps(
@@ -50,14 +41,13 @@ def claim(token, facility, client_ip=None):
         "token": token,
     }
 
-    # sign the client/server nonce combo to attach to the request
-    message = "{client_nonce}:{server_nonce}".format(**data)
+    # sign the server nonce to attach to the request
+    message = "{server_nonce}".format(**data)
     data["signature"] = client_cert.sign(message)
 
     # attempt to claim the facility
     response = requests.post(
-        urljoin(PORTAL_URL, "portal/api/public/v1/claim/"), data=data
+        urljoin(PORTAL_URL, "portal/api/public/v1/registerfacility/"), data=data
     )
-    Nonce.use_nonce(client_nonce.id)
     response.raise_for_status()
     return response
